Collect all parts of the chat page in `ChatsPage.kt`
